give actors the gift of movement.

+++ include/badger/position.h
@@ -6,8 +6,8 @@ namespace Badger {
   struct Position {
 
     // The x and y coordinates of the object's footprint in the world.
-    unsigned int x;
-    unsigned int y;
+    double x;
+    double y;
 
     // The width and height of the object's footprint in the world.
     unsigned int width;
